+++ plugins/postcss-color-mod-function/lib/color.js
@@ -1,11 +1,11 @@
-import number from 'big.js';
+import { convertRGBtoHSL, convertRGBtoHWB, convertHSLtoRGB, convertHSLtoHWB, convertHWBtoRGB, convertHWBtoHSL, convertRGBtoH } from './conversions';
 
 export default class Color {
 	constructor(color) {
 		this.color = Object(Object(color).color || color);
 
 		if (color.colorspace === 'rgb') {
-			this.color.hue = rgb2hue(color.red, color.green, color.blue, color.hue || 0);
+			this.color.hue = convertRGBtoH(color.red, color.green, color.blue, color.hue || 0);
 		}
 	}
 
@@ -133,6 +133,10 @@ export default class Color {
 		return color2hwbString(this.color);
 	}
 
+	toLegacy() {
+		return color2legacyString(this.color);
+	}
+
 	toRGB() {
 		return color2rgbString(this.color);
 	}
@@ -149,20 +153,20 @@ export default class Color {
 /* Blending
 /* ========================================================================== */
 
-function blend(base, color, percentage, colorspace, isBlendingAlpha) { // eslint-disable-line max-params
-	const addition    = number(percentage).div(100);
-	const subtraction = number(1).minus(addition);
+function blend(base, color, percentage, colorspace, isBlendingAlpha) {
+	const addition    = percentage / 100;
+	const subtraction = 1 - addition;
 
 	if (colorspace === 'hsl') {
 		const { hue: h1, saturation: s1, lightness: l1, alpha: a1 } = color2hsl(base);
 		const { hue: h2, saturation: s2, lightness: l2, alpha: a2 } = color2hsl(color);
 
 		const [hue, saturation, lightness, alpha] = [
-			Number(number(h1).times(subtraction).plus(number(h2).times(addition))),
-			Number(number(s1).times(subtraction).plus(number(s2).times(addition))),
-			Number(number(l1).times(subtraction).plus(number(l2).times(addition))),
+			h1 * subtraction + h2 * addition,
+			s1 * subtraction + s2 * addition,
+			l1 * subtraction + l2 * addition,
 			isBlendingAlpha
-				? Number(number(a1).times(subtraction).plus(number(a2).times(addition)))
+				? a1 * subtraction + a2 * addition
 			: a1
 		];
 
@@ -172,10 +176,12 @@ function blend(base, color, percentage, colorspace, isBlendingAlpha) { // eslint
 		const { hue: h2, whiteness: w2, blackness: b2, alpha: a2 } = color2hwb(color);
 
 		const [hue, whiteness, blackness, alpha] = [
-			Number(number(h1).times(subtraction).plus(number(h2).times(addition))),
-			Number(number(w1).times(subtraction).plus(number(w2).times(addition))),
-			Number(number(b1).times(subtraction).plus(number(b2).times(addition))),
-			isBlendingAlpha ? Number(number(a1).times(subtraction).plus(number(a2).times(addition))) : a1
+			h1 * subtraction + h2 * addition,
+			w1 * subtraction + w2 * addition,
+			b1 * subtraction + b2 * addition,
+			isBlendingAlpha
+				? a1 * subtraction + a2 * addition
+			: a1
 		];
 
 		return { hue, whiteness, blackness, alpha, colorspace: 'hwb' };
@@ -184,10 +190,12 @@ function blend(base, color, percentage, colorspace, isBlendingAlpha) { // eslint
 		const { red: r2, green: g2, blue: b2, alpha: a2 } = color2rgb(color);
 
 		const [red, green, blue, alpha] = [
-			Number(number(r1).times(subtraction).plus(number(r2).times(addition))),
-			Number(number(g1).times(subtraction).plus(number(g2).times(addition))),
-			Number(number(b1).times(subtraction).plus(number(b2).times(addition))),
-			isBlendingAlpha ? Number(number(a1).times(subtraction).plus(number(a2).times(addition))) : a1
+			r1 * subtraction + r2 * addition,
+			g1 * subtraction + g2 * addition,
+			b1 * subtraction + b2 * addition,
+			isBlendingAlpha
+				? a1 * subtraction + a2 * addition
+			: a1
 		];
 
 		return { red, green, blue, alpha, colorspace: 'rgb' };
@@ -214,7 +222,7 @@ function assign(base, channels) {
 
 			if (isRGB) {
 				// conditionally preserve the hue
-				color.hue = rgb2hue(color.red, color.green, color.blue, base.hue || 0);
+				color.hue = convertRGBtoH(color.red, color.green, color.blue, base.hue || 0);
 			}
 		}
 	);
@@ -230,183 +238,44 @@ function normalize(value, channel) {
 	const min = 0;
 	const max = isHue ? 360 : 100;
 
-	const modifiedValue = Math.min(Math.max(isHue
-		? number(value).mod(360)
+	const normalizedValue = Math.min(Math.max(isHue
+		? value % 360
 	: value, min), max);
 
-	return modifiedValue
+	return normalizedValue;
 }
 
 /* Convert colors
 /* ========================================================================== */
 
-function color2hsl(color) {
-	return color.colorspace === 'rgb'
-		? rgb2hsl(color, color.hue)
+function color2rgb(color) {
+	const [ red, green, blue ] = color.colorspace === 'hsl'
+		? convertHSLtoRGB(color.hue, color.saturation, color.lightness)
 	: color.colorspace === 'hwb'
-		? rgb2hsl(hwb2rgb(color), color.hue)
-	: color;
+		? convertHWBtoRGB(color.hue, color.whiteness, color.blackness)
+	: [ color.red, color.green, color.blue ];
+
+	return { red, green, blue, hue: color.hue, alpha: color.alpha, colorspace: 'rgb' };
+}
+
+function color2hsl(color) {
+	const [ hue, saturation, lightness ] = color.colorspace === 'rgb'
+		? convertRGBtoHSL(color.red, color.green, color.blue, color.hue)
+	: color.colorspace === 'hwb'
+		? convertHWBtoHSL(color.hue, color.whiteness, color.blackness)
+	: [ color.hue, color.saturation, color.lightness ];
+
+	return { hue, saturation, lightness, alpha: color.alpha, colorspace: 'hsl' };
 }
 
 function color2hwb(color) {
-	return color.colorspace === 'rgb'
-		? rgb2hwb(color, color.hue)
+	const [ hue, whiteness, blackness ] = color.colorspace === 'rgb'
+		? convertRGBtoHWB(color.red, color.green, color.blue, color.hue)
 	: color.colorspace === 'hsl'
-		? rgb2hwb(hsl2rgb(color), color.hue)
-	: color;
-}
+		? convertHSLtoHWB(color.hue, color.saturation, color.lightness)
+	: [ color.hue, color.whiteness, color.blackness ];
 
-function color2rgb(color) {
-	return color.colorspace === 'hsl'
-		? hsl2rgb(color)
-	: color.colorspace === 'hwb'
-		? hwb2rgb(color)
-	: color;
-}
-
-/* Convert HSL to RGB
-/* ========================================================================== */
-
-function hsl2rgb({ hue, saturation, lightness, alpha = 100 }) {
-	const t2 = lightness <= 50
-		? Number(
-			number(lightness).div(100).times(
-				number(saturation).div(100).plus(1)
-			)
-		)
-	: Number(
-		number(lightness).div(100).plus(
-			number(saturation).div(100)
-		).minus(
-			number(lightness).div(100).times(
-				number(saturation).div(100)
-			)
-		)
-	);
-
-	const t1 = Number(number(lightness).div(100).times(2).minus(t2));
-
-	const red   = hue2channel(t1, t2, Number(number(hue).div(60).plus(2)));
-	const green = hue2channel(t1, t2, Number(number(hue).div(60)));
-	const blue  = hue2channel(t1, t2, Number(number(hue).div(60).minus(2)));
-
-	return { hue, red, green, blue, alpha, colorspace: 'rgb' };
-}
-
-/* Convert HWB to RGB
-/* ========================================================================== */
-
-function hwb2rgb({ hue, whiteness, blackness, alpha = 100 }) {
-	const ratio = Number(number(whiteness).plus(blackness));
-	const rwhiteness = ratio > 100 ? Number(number(whiteness).div(ratio)) : whiteness;
-	const rblackness = ratio > 100 ? Number(number(blackness).div(ratio)) : blackness;
-	const value = Number(number(100).minus(rblackness));
-	const hexagon = number(6).times(hue).div(360);
-
-	const hexagonFloor = Math.floor(hexagon);
-
-	const hexagonF = hexagonFloor % 6
-		? number(1).minus(
-			hexagon.minus(hexagonFloor)
-		)
-	: hexagon.minus(hexagonFloor);
-
-	const interpolation = Number(number(rwhiteness).plus(
-		hexagonF.times(
-			number(value).minus(rwhiteness)
-		)
-	));
-
-	const [red, green, blue] = hexagonFloor % 6 === 5
-		? [value, rwhiteness, interpolation]
-	: hexagonFloor % 6 === 4
-		? [interpolation, rwhiteness, value]
-	: hexagonFloor % 6 === 3
-		? [rwhiteness, interpolation, value]
-	: hexagonFloor % 6 === 2
-		? [rwhiteness, value, interpolation]
-	: hexagonFloor % 6 === 1
-		? [interpolation, value, rwhiteness]
-	: [value, interpolation, rwhiteness];
-
-	return { hue, red, green, blue, alpha, colorspace: 'rgb' };
-}
-
-/* Convert RGB to HSL
-/* ========================================================================== */
-
-function rgb2hsl({ red, green, blue, alpha = 100 }, fallback = 0) { // eslint-disable-line max-params
-	const hue        = rgb2hue(red, green, blue, fallback);
-	const whiteness  = rgb2whiteness(red, green, blue);
-	const value      = rgb2value(red, green, blue);
-	const lightness  = wv2lightness(whiteness, value);
-	const saturation = lvw2saturation(lightness, value, whiteness);
-
-	return { hue, saturation, lightness, alpha, colorspace: 'hsl' };
-}
-
-/* Convert RGB to HWB
-/* ========================================================================== */
-
-function rgb2hwb({ red, green, blue, alpha = 100 }, fallback = 0) { // eslint-disable-line max-params
-	const hue       = rgb2hue(red, green, blue, fallback);
-	const whiteness = rgb2whiteness(red, green, blue);
-	const value     = rgb2value(red, green, blue);
-	const blackness = Number(number(100).minus(value));
-
-	return { hue, whiteness, blackness, alpha, colorspace: 'hwb' };
-}
-
-/* Convert Hue to RGB
-/* ========================================================================== */
-
-function hue2channel(t1, t2, hue) {
-	const huerange = hue < 0
-		? number(hue).plus(6)
-	: hue >= 6
-		? number(hue).minus(6)
-	: number(hue);
-
-	const rgb = huerange < 1
-		? Number(number(number(t2).minus(t1)).times(hue).plus(t1))
-	: hue < 3
-		? t2
-	: hue < 4
-		? Number(number(number(t2).minus(t1)).times(number(4).minus(hue)).plus(t1))
-	: t1;
-
-	return Number(number(rgb).times(100));
-}
-
-/* Convert RGB to Hue
-/* ========================================================================== */
-
-function rgb2hue(red, green, blue, fallback) { // eslint-disable-line max-params
-	const whiteness = rgb2whiteness(red, green, blue);
-	const value     = rgb2value(red, green, blue);
-	const chroma    = vw2chroma(value, whiteness);
-
-	if (chroma === 0) {
-		return fallback;
-	} else {
-		const segment = value === red
-			? number(green).minus(blue).div(chroma)
-		: value === green
-			? number(blue).minus(red).div(chroma)
-		: number(red).minus(green).div(chroma);
-
-		const shift = value === red
-			? segment < 0
-				? 360 / 60
-				: 0 / 60
-		: value === green
-			? 120 / 60
-		: 240 / 60;
-
-		const hue = Number(number(segment).plus(shift).times(60));
-
-		return hue;
-	}
+	return { hue, whiteness, blackness, alpha: color.alpha, colorspace: 'hwb' };
 }
 
 /* Contrast functions
@@ -415,9 +284,10 @@ function rgb2hue(red, green, blue, fallback) { // eslint-disable-line max-params
 function contrast(color, percentage) {
 	// https://drafts.csswg.org/css-color/#contrast-adjuster
 	const hwb = color2hwb(color);
+	const rgb = color2rgb(color);
 
 	// compute the luminance of the color.
-	const luminance = rgb2luminance(color.red, color.green, color.blue);
+	const luminance = rgb2luminance(rgb.red, rgb.green, rgb.blue);
 
 	// the maximum-contrast color, if it is less than .5
 	const maxContrastColor = luminance < 0.5
@@ -435,7 +305,7 @@ function contrast(color, percentage) {
 	// otherwise, the maximum-contrast color
 	: maxContrastColor;
 
-	// color(maximum-contrast blend(minimum-contrast <percentage> hwb))
+	// color(maximum-contrast blend(minimum-contrast <percentage> hwb)));
 	return blend(maxContrastColor, minContrastColor, percentage, 'hwb', false);
 }
 
@@ -448,9 +318,9 @@ function colors2contrast(color1, color2) {
 
 	return l1 > l2
 		// if l1 is the relative luminance of the lighter of the colors
-		? Number(number(l1).plus(0.05).div(number(l2).plus(0.05)))
+		? (l1 + 0.05) / (l2 + 0.05)
 	// otherwise, if l2 is the relative luminance of the lighter of the colors
-	: Number(number(l2).plus(0.05).div(number(l2).plus(0.05)));
+	: (l2 + 0.05) / (l1 + 0.05);
 }
 
 function rgb2luminance(red, green, blue) {
@@ -461,14 +331,14 @@ function rgb2luminance(red, green, blue) {
 	];
 
 	// https://drafts.csswg.org/css-color/#luminance
-	const luminance = Number(number(0.2126).times(redLuminance).plus(number(0.7152).times(greenLuminance)).plus(number(0.0722).times(blueLuminance)));
+	const luminance = 0.2126 * redLuminance + 0.7152 * greenLuminance + 0.0722 * blueLuminance;
 
 	return luminance;
 }
 
 function channel2luminance(value) {
 	// https://drafts.csswg.org/css-color/#luminance
-	const luminance = value <= 0.03928 ? Number(number(value).div(12.92)) : Math.pow(number(value).plus(0.055).div(1.055), 2.4);
+	const luminance = value <= 0.03928 ? value / 12.92 : Math.pow((value + 0.055) /1.055, 2.4);
 
 	return luminance;
 }
@@ -478,18 +348,18 @@ function colors2contrastRatioColor(hwb, maxHWB) {
 	const modifiedHWB = Object.assign({}, hwb);
 
 	// values to be used for linear interpolations in HWB space
-	let minW = number(hwb.whiteness);
-	let minB = number(hwb.blackness);
-	let maxW = number(maxHWB.whiteness);
-	let maxB = number(maxHWB.blackness);
+	let minW = hwb.whiteness;
+	let minB = hwb.blackness;
+	let maxW = maxHWB.whiteness;
+	let maxB = maxHWB.blackness;
 
 	// find the color with the smallest contrast ratio with the base color that is greater than 4.5
-	while (Number(minW.minus(maxW).abs()) > 100 || Number(minB.minus(maxB).abs()) > 100) {
-		const midW = maxW.plus(minW).div(2).round();
-		const midB = maxB.plus(minB).div(2).round();
+	while (Math.abs(minW - maxW) > 100 || Math.abs(minB - maxB) > 100) {
+		const midW = Math.round((maxW + minW) / 2);
+		const midB = Math.round((maxB + minB) / 2);
 
-		modifiedHWB.whiteness = Number(midW);
-		modifiedHWB.blackness = Number(midB);
+		modifiedHWB.whiteness = midW;
+		modifiedHWB.blackness = midB;
 
 		if (colors2contrast(modifiedHWB, hwb) > 4.5) {
 			maxW = midW;
@@ -501,45 +371,6 @@ function colors2contrastRatioColor(hwb, maxHWB) {
 	}
 
 	return modifiedHWB;
-}
-
-/* Convert RGB to Whiteness
-/* ========================================================================== */
-
-function rgb2whiteness(red, green, blue) {
-	return Math.min(red, green, blue);
-}
-
-/* Convert RGB to Value
-/* ========================================================================== */
-
-function rgb2value(red, green, blue) {
-	return Math.max(red, green, blue)
-}
-
-/* Convert Whiteness and Value to Lightness
-/* ========================================================================== */
-
-function wv2lightness(whiteness, value) {
-	return Number(number(whiteness).plus(value).div(2));
-}
-
-/* Convert Value and Whiteness to Chroma
-/* ========================================================================== */
-
-function vw2chroma(value, whiteness) {
-	return Number(number(value).minus(whiteness));
-}
-
-/* Convert Lightness, Value, and Whiteness to Saturation
-/* ========================================================================== */
-
-function lvw2saturation(lightness, value, whiteness) {
-	return whiteness === value
-		? 0
-	: lightness < 50
-		? Number(number(value).minus(whiteness).div(number(value).plus(whiteness)).times(100))
-	: Number(number(value).minus(whiteness).div(number(200).minus(value).minus(whiteness)).times(100));
 }
 
 /* Match
@@ -562,9 +393,9 @@ function color2hslString(color) {
 	const hsl        = color2hsl(color);
 	const isOpaque   = hsl.alpha === 100;
 	const hue        = hsl.hue;
-	const saturation = number(hsl.saturation).round(10);
-	const lightness  = number(hsl.lightness).round(10);
-	const alpha      = number(hsl.alpha).round(10);
+	const saturation = Math.round(hsl.saturation * 10000000000) / 10000000000;
+	const lightness  = Math.round(hsl.lightness * 10000000000) / 10000000000;
+	const alpha      = Math.round(hsl.alpha * 10000000000) / 10000000000;
 
 	return `hsl(${hue} ${saturation}% ${lightness}%${isOpaque
 		? ''
@@ -575,9 +406,9 @@ function color2hwbString(color) {
 	const hwb       = color2hwb(color);
 	const isOpaque  = hwb.alpha === 100;
 	const hue       = hwb.hue;
-	const whiteness = number(hwb.whiteness).round(10);
-	const blackness = number(hwb.blackness).round(10);
-	const alpha     = number(hwb.alpha).round(10);
+	const whiteness = Math.round(hwb.whiteness * 10000000000) / 10000000000;
+	const blackness = Math.round(hwb.blackness * 10000000000) / 10000000000;
+	const alpha     = Math.round(hwb.alpha * 10000000000) / 10000000000;
 
 	return `hwb(${hue} ${whiteness}% ${blackness}%${isOpaque
 		? ''
@@ -587,26 +418,46 @@ function color2hwbString(color) {
 function color2rgbString(color) {
 	const rgb      = color2rgb(color);
 	const isOpaque = rgb.alpha === 100;
-	const red      = number(rgb.red).round(10);
-	const green    = number(rgb.green).round(10);
-	const blue     = number(rgb.blue).round(10);
-	const alpha    = number(rgb.alpha).round(10);
+	const red      = Math.round(rgb.red * 10000000000) / 10000000000;
+	const green    = Math.round(rgb.green * 10000000000) / 10000000000;
+	const blue     = Math.round(rgb.blue * 10000000000) / 10000000000;
+	const alpha    = Math.round(rgb.alpha * 10000000000) / 10000000000;
 
 	return `rgb(${red}% ${green}% ${blue}%${isOpaque
 		? ''
 	: ` / ${alpha}%`})`;
 }
 
+function color2legacyString(color) {
+	return color.colorspace === 'hsl'
+		? color2hslLegacyString(color)
+	: color2rgbLegacyString(color);
+}
+
 function color2rgbLegacyString(color) {
 	const rgb      = color2rgb(color);
 	const isOpaque = rgb.alpha === 100;
 	const name     = isOpaque ? 'rgb' : 'rgba';
-	const red      = number(rgb.red).times(2.55).round(0);
-	const green    = number(rgb.green).times(2.55).round(0);
-	const blue     = number(rgb.blue).times(2.55).round(0);
-	const alpha    = number(rgb.alpha).div(100).round(10);
+	const red      = Math.round(rgb.red * 255 / 100);
+	const green    = Math.round(rgb.green * 255 / 100);
+	const blue     = Math.round(rgb.blue * 255 / 100);
+	const alpha    = Math.round(rgb.alpha / 100 * 10000000000) / 10000000000;
 
 	return `${name}(${red}, ${green}, ${blue}${isOpaque
+		? ''
+	: `, ${alpha}`})`;
+}
+
+function color2hslLegacyString(color) {
+	const hsl        = color2hsl(color);
+	const isOpaque   = hsl.alpha === 100;
+	const name       = isOpaque ? 'hsl' : 'hsla';
+	const hue        = hsl.hue;
+	const saturation = Math.round(hsl.saturation * 10000000000) / 10000000000;
+	const lightness  = Math.round(hsl.lightness * 10000000000) / 10000000000;
+	const alpha      = Math.round(hsl.alpha / 100 * 10000000000) / 10000000000;
+
+	return `${name}(${hue}, ${saturation}%, ${lightness}%${isOpaque
 		? ''
 	: `, ${alpha}`})`;
 }