+++ plugins/postcss-color-mod-function/lib/manage-unresolved.js
@@ -1,4 +1,4 @@
-export default function manageUnresolved(node, opts, word, message) { // eslint-disable-line max-params
+export default function manageUnresolved(node, opts, word, message) {
 	if ('warn' === opts.unresolved) {
 		opts.decl.warn(opts.result, message, { word });
 	} else if ('ignore' !== opts.unresolved) {