+++ plugins/postcss-color-mod-function/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes to PostCSS color-mod() Function
 
+### 2.1.0 (January 20, 2018)
+
+- Added support for legacy (comma-separated) `hsl()` colors
+- Added support for all `<hue>` units
+- Added use of legacy (comma-separated) `hsl()` colors when appropriate
+- Improved color conversions
+- Improved support for all `rgb()` colors
+- Removed external math and color dependencies
+
 ### 2.0.0 (January 17, 2018)
 
 - Reverse blend/blenda percentage calculations (breaking change)